requirements.txt
Django==4.2.7
django-ninja==1.0.1
psycopg2-binary==2.9.9
APScheduler==3.10.4
python-decouple==3.8
django-cors-headers==4.3.1
Pillow==10.1.0